no relative paths for input folders; undo #8
