+++ interfaces/final-object.interface.ts
@@ -16,7 +16,7 @@ export interface FinalObject {
   addTags: string[];             // tags to add
   hubName: string;               // the name of the hub -- for recently-opened
   images: ImageElement[];
-  // inputDir: string;           // became `inputDirs` in VHA3
+  // inputDir: string;           // became `inputDirs` in VHA3 - gets upgraded to `inputDirs` on first open in VHA3
   inputDirs: InputSources;       // map the `inputSource` number to input directory (replaces `inputDir`)
   numOfFolders: number;          // number of folders - is always re-counted when app starts
   removeTags: string[];          // tags to remove
